+++ src/view/_components/JSONEditor/JSONEditor.ts
@@ -2,10 +2,10 @@ import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application';
 
 import type { CombinedSvelteApplicationOptions, ConstructorApplicationOptions } from 'src/extensions';
 import { ErrorMsg } from 'src/utils';
-import { type Writable, get } from 'svelte/store';
+import { type Writable, get, writable } from 'svelte/store';
 import BasicAppShell from './JSONEditor.svelte';
 
-interface extra { store: Writable<object>; permission: boolean; stasis: Writable<boolean> }
+interface extra { store: Writable<object>; permission?: boolean; readOnly?: true; stasis?: Writable<boolean> }
 export default class JSONEditorApp extends SvelteApplication {
 	constructor(_options: ConstructorApplicationOptions & extra) {
 		// @ts-expect-error TJS-2-TS
@@ -30,15 +30,15 @@ export default class JSONEditorApp extends SvelteApplication {
 					const store = this.options.store;
 					if (!store) throw new ErrorMsg('No Store Provided in ItemAnimationsApp!');
 					// @ts-expect-error TJS-2-TS
-					const permission = this.options.permission ?? false;
+					const readOnly = this.options.readOnly ?? !this.options.permission;
 					// @ts-expect-error TJS-2-TS
-					const stasis = this.options.stasis;
+					const stasis = this.options.stasis ?? writable(false);
 					return {
 						store,
-						permission,
+						readOnly,
 						stasis,
 					};
-				} as () => { store: Writable<object>; permission: boolean },
+				} as () => { store: Writable<object>; readOnly: boolean; stasis: Writable<boolean> },
 			},
 		});
 	}