+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](<https://semver.org/spec/v2.0.
 
 ### Added
 
-- Setting overrides to Animations Editor.
+- Live preview mode to JSON Editor. Accessed by Shift clicking.
 - `action` trigger. Activates animations whenever an action is posted, like Recognize Spell or Change Shape.
 - Animation History menu. Stores animations played in the current session. Removed on refresh.
 - Delay Animations settings. ([#111](<https://github.com/MrVauxs/pf2e-graphics/issues/111>))
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](<https://semver.org/spec/v2.0.
 - Animation volume adjustment in both module settings and playlist tab. ([#89](<https://github.com/MrVauxs/pf2e-graphics/issues/89>))
 - Holding Shift to see processed version of premade animations in Preset Animations menu.
 - Ability to see errorenous sounds in Preset Animations menu.
-- Ability to use multiple Triggers for animations.
+- Support for multiple Triggers in animations. Not supported by the
 - Warnings for when you dont have JB2A enabled.
 - Conditions triggering `toggle` animations.
 - **New Sounds**